build fix
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@31839 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMGuruMeditation.cpp
@@ -125,7 +125,7 @@ static DECLCALLBACK(void) vmmR3FatalDumpInfoHlp_pfnPrintfV(PCDBGFINFOHLP pHlp, c
     }
     if (pMyHlp->fRecSummary)
     {
-        uint32_t cchLeft = sizeof(pMyHlp->szSummary) - pMyHlp->offSummary;
+        size_t cchLeft = sizeof(pMyHlp->szSummary) - pMyHlp->offSummary;
         if (cchLeft > 1)
         {
             va_list args2;
